refactor(gemini): move browser adapter into provider package
