+++ tests/test_gemini_model_selection.py
@@ -2,7 +2,7 @@ import pytest
 import asyncio
 from unittest.mock import MagicMock, AsyncMock, patch
 from fastapi import HTTPException
-from app.services.browser.adapters.gemini_adapter import GeminiProviderAdapter
+from app.services.providers.gemini.browser_adapter import GeminiProviderAdapter
 from app.services.providers.gemini.playwright_adapter import (
     GeminiPlaywrightAdapter, 
     PlaywrightRequestState, 
@@ -88,7 +88,7 @@ async def test_find_model_picker_fallbacks(adapter, mock_page):
     mock_fallback_loc.first = mock_fallback_el
     
     def locator_side_effect(selector):
-        from app.services.browser.adapters.scripts.gemini_scripts import SELECTORS
+        from app.services.providers.gemini.scripts.gemini_scripts import SELECTORS
         if selector == SELECTORS["MODEL_PICKER"]:
             return mock_primary
         if selector == 'button[aria-label*="Select model"]':
@@ -141,7 +141,7 @@ async def test_select_model_success(adapter, mock_page):
     
     # Route locator calls
     def locator_side_effect(selector):
-        from app.services.browser.adapters.scripts.gemini_scripts import SELECTORS
+        from app.services.providers.gemini.scripts.gemini_scripts import SELECTORS
         if selector == SELECTORS["MODEL_PICKER"]:
             return mock_picker_loc
         if selector == SELECTORS["MODEL_OPTION"]:
@@ -198,7 +198,7 @@ async def test_select_model_fails_when_gated_advanced(adapter, mock_page):
     
     # Route locator calls
     def locator_side_effect(selector):
-        from app.services.browser.adapters.scripts.gemini_scripts import SELECTORS
+        from app.services.providers.gemini.scripts.gemini_scripts import SELECTORS
         if selector == SELECTORS["MODEL_PICKER"]:
             return mock_picker_loc
         if selector == SELECTORS["MODEL_OPTION"]:
@@ -247,7 +247,7 @@ async def test_select_model_collision_prevention_flash_vs_lite(adapter, mock_pag
     
     # Route locator calls
     def locator_side_effect(selector):
-        from app.services.browser.adapters.scripts.gemini_scripts import SELECTORS
+        from app.services.providers.gemini.scripts.gemini_scripts import SELECTORS
         if selector == SELECTORS["MODEL_PICKER"]:
             return mock_picker_loc
         if selector == SELECTORS["MODEL_OPTION"]:
